backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "A sample fastapi project created with create-fastapi-project"
authors = [ "Your Name <your@email.com>",]
readme = "README.md"
[[tool.poetry.packages]]
include = "app"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
mypy = "^1.10.0"
ruff = "^0.4.8"
black = "^24.4.2"

[tool.black]
line-length = 88
target-version = [ "py37", "py38", "py39", "py310", "py311",]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.ruff]
line-length = 88
exclude = [ ".git", "__pycache__", ".mypy_cache", ".pytest_cache",]
select = [ "E", "W", "F", "C", "B",]
ignore = [ "B904", "B006", "E501", "B008", "C901",]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [ "alembic", "__pycache__",]

[tool.build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.25"}
asyncer = "^0.0.7"
httpx = "^0.27.0"
requests = "^2.32.3"
msal = "^1.28.1"
msgraph-sdk = "^1.4.0"
azure-identity = "^1.16.1"

[tool.ruff.per-file-ignores]
"__init__.py" = [ "F401",]
